requirements.txt
Django>=4.0
django-guardian>=2.4.0
djangorestframework>=3.14.0
